Rename `spectra` module to `converters` to better describe its contents
